Сделана валидация через посредник
